rb3-builtin-cameras/requirements.txt
numpy==2.3.3
PyGObject==3.50.0
Pillow==11.3.0
